# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Backlite API
  version: "1.0.0"
  description: >
    Backlite is a self-hostable Backlite fork that runs coding agents (Claude
    Code or Codex) in ephemeral local Docker containers. Tasks come in via REST
    API; the orchestrator runs agents, persists artifacts, and cleans up.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/v1/readings:
    get:
      operationId: listReadings
      summary: List saved readings
      description: Returns saved reading results newest-first with offset pagination. Requires bearer authentication when API keys are configured.
      parameters:
        - name: limit
          in: query
          description: Maximum number of readings to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of readings to skip for pagination
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Page of saved readings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingListEnvelope"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: API key lacks required scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      security:
        - BearerAuth: [readings:read]
  /api/v1/readings/{id}:
    get:
      operationId: getReading
      summary: Get a saved reading
      description: Returns one saved reading result by ID. Requires bearer authentication when API keys are configured.
      parameters:
        - name: id
          in: path
          required: true
          description: Reading ID
          schema:
            type: string
      responses:
        "200":
          description: Reading details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingEnvelope"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: API key lacks required scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Reading not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      security:
        - BearerAuth: [readings:read]
  /api/v1/tasks:
    post:
      operationId: createTask
      summary: Create a new task
      description: >
        Creates a coding or review task. Prompt is the only required field —
        the agent container's Prep stage infers repo_url, target_branch, and
        task_mode from the prompt. If the prompt starts with a GitHub PR URL,
        review mode is auto-detected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskEnvelope"
          links:
            GetTask:
              operationId: getTask
              parameters:
                id: $response.body#/data/id
            CancelTask:
              operationId: cancelTask
              parameters:
                id: $response.body#/data/id
            GetTaskLogs:
              operationId: getTaskLogs
              parameters:
                id: $response.body#/data/id
        "400":
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: API key lacks required scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal server error (e.g. database failure)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      security:
        - BearerAuth: [tasks:write]

    get:
      operationId: listTasks
      summary: List tasks
      description: Returns a paginated list of tasks, optionally filtered by status.
      parameters:
        - name: status
          in: query
          description: Filter tasks by status
          required: false
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: limit
          in: query
          description: Maximum number of tasks to return
          required: false
          schema:
            type: integer
            minimum: 1
        - name: offset
          in: query
          description: Number of tasks to skip for pagination
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: List of tasks (empty array when none match)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListEnvelope"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: API key lacks required scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      security:
        - BearerAuth: [tasks:read]
  /api/v1/tasks/{id}:
    get:
      operationId: getTask
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID (bf_-prefixed ULID)
          schema:
            type: string
            pattern: "^bf_[0-9A-Z]{26}$"
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskEnvelope"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: API key lacks required scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      security:
        - BearerAuth: [tasks:read]

    delete:
      operationId: cancelTask
      summary: Cancel or delete a task
      description: >
        If the task is running, provisioning, or recovering, marks it as
        cancelled (the orchestrator will stop the container). If the task is
        pending, deletes it immediately. Terminal tasks are a no-op.
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID (bf_-prefixed ULID)
          schema:
            type: string
            pattern: "^bf_[0-9A-Z]{26}$"
      responses:
        "204":
          description: Task cancelled or deleted (no body)
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: API key lacks required scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      security:
        - BearerAuth: [tasks:write]
  /api/v1/tasks/{id}/retry:
    post:
      operationId: retryTask
      summary: Retry a failed, interrupted, or cancelled task
      description: >
        Atomically requeues a task for retry. The task must be in a terminal
        retryable state (failed, interrupted, or cancelled), be marked ready
        for retry (container cleanup complete), and be under the configured
        user retry cap. Returns 409 Conflict if the task is not retryable.
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID (bf_-prefixed ULID)
          schema:
            type: string
            pattern: "^bf_[0-9A-Z]{26}$"
      responses:
        "200":
          description: Task successfully requeued for retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskEnvelope"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: API key lacks required scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Task cannot be retried (wrong status, not ready, or cap reached)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      security:
        - BearerAuth: [tasks:write]
  /api/v1/tasks/{id}/logs:
    get:
      operationId: getTaskLogs
      summary: Get task container logs
      description: >
        Streams container logs for a running task. If the task has not yet
        started a container, returns a status/error text message instead.
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID (bf_-prefixed ULID)
          schema:
            type: string
            pattern: "^bf_[0-9A-Z]{26}$"
        - name: tail
          in: query
          description: Number of log lines to return from the end
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Container log output (or status message if not yet running)
          content:
            text/plain:
              schema:
                type: string
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: API key lacks required scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "502":
          description: Failed to fetch logs from the container runtime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      security:
        - BearerAuth: [tasks:read]
  /api/v1/tasks/{id}/output:
    get:
      operationId: getTaskOutput
      summary: Get persisted agent output log
      description: >
        Returns the agent's stdout log (container_output.log) written by the
        orchestrator to the configured data directory after the task's
        container exits. Returns 404 if the task has not finished, if
        save_agent_output was disabled, or if the file has not yet been
        written.
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID (bf_-prefixed ULID)
          schema:
            type: string
            pattern: "^bf_[0-9A-Z]{26}$"
      responses:
        "200":
          description: Agent output log
          content:
            text/plain:
              schema:
                type: string
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: API key lacks required scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Output not found (task missing, still running, or output gated off)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      security:
        - BearerAuth: [tasks:read]
  /api/v1/tasks/{id}/output.json:
    get:
      operationId: getTaskOutputJSON
      summary: Get persisted task metadata snapshot
      description: >
        Returns the JSON task metadata snapshot (task.json) written to the
        configured data directory alongside the agent output log after the
        task's container exits.
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID (bf_-prefixed ULID)
          schema:
            type: string
            pattern: "^bf_[0-9A-Z]{26}$"
      responses:
        "200":
          description: Task metadata snapshot
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: API key lacks required scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Output not found (task missing, still running, or output gated off)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      security:
        - BearerAuth: [tasks:read]
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "task not found"
    TaskEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Task"
    TaskListEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Task"
    ReadingListEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/ReadingPage"
    ReadingEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Reading"
    ReadingPage:
      type: object
      required: [readings, limit, offset, has_more]
      properties:
        readings:
          type: array
          items:
            $ref: "#/components/schemas/Reading"
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        offset:
          type: integer
          minimum: 0
          example: 0
        has_more:
          type: boolean
          example: false

    # -------------------------------------------------------------------------
    # Core types
    # -------------------------------------------------------------------------
    TaskStatus:
      type: string
      enum:
        - pending
        - provisioning
        - running
        - completed
        - failed
        - interrupted
        - cancelled
        - recovering
      example: pending
...
    CreateTaskRequest:
      type: object
      additionalProperties: false
      description: >
        Request body for creating a task. Prompt is the only required field —
        the agent container's Prep stage infers repo_url, target_branch, and
        task_mode from the prompt.
      required: [prompt]
      properties:
        prompt:
          type: string
          description: Task instructions for the agent. For read-mode tasks, the URL to summarize.
          example: "Fix the authentication bug in https://github.com/owner/repo"
        task_mode:
          type: string
          enum: [auto, read]
          description: >
            Explicit task mode. Omit (or "auto") to let the agent's Prep stage infer
            code vs. review from the prompt. Set to "read" to create a reading-mode
            task (the prompt is the URL to summarize). Explicit "code" and "review"
            are not accepted — they are always inferred.
          example: read
        force:
          type: boolean
          description: For read-mode tasks, skip duplicate detection and upsert on completion. No effect for other modes.
          example: false
        harness:
          type: string
          enum: [claude_code, codex]
          description: Agent harness to use. Defaults to server configuration.
          example: claude_code
        context:
          type: string
          description: Additional context provided to the agent.
          example: "This service uses chi router and SQLite via database/sql"
        model:
          type: string
          description: LLM model override. Defaults to server configuration.
          example: "claude-sonnet-4-6"
        effort:
          type: string
          enum: [low, medium, high, xhigh]
          description: Agent effort level. Defaults to server configuration.
          example: medium
        max_budget_usd:
          type: number
          minimum: 0
          description: Maximum spend in USD. Defaults to server configuration.
          example: 5.0
        max_runtime_sec:
          type: integer
          minimum: 0
          description: Maximum runtime in seconds. Defaults to server configuration.
          example: 1800
        max_turns:
          type: integer
          minimum: 0
          description: Maximum agent turns. Defaults to server configuration.
          example: 100
        create_pr:
          type: boolean
          description: Whether to open a pull request when done. Defaults to server configuration.
          example: true
        self_review:
          type: boolean
          description: Whether the agent should self-review its output. Defaults to server configuration.
          example: false
        save_agent_output:
          type: boolean
          description: Whether to persist agent output for retrieval via the output endpoints. Defaults to server configuration.
          example: true
        pr_title:
          type: string
          description: Custom pull request title. Defaults to agent-generated title.
          example: "Fix authentication bug"
        pr_body:
          type: string
          description: Custom pull request body. Defaults to agent-generated body.
          example: "This PR fixes the authentication bug by..."
        allowed_tools:
          type: array
          items:
            type: string
          description: Restrict agent to a specific set of tools.
          example: ["Bash", "Edit", "Read"]
        claude_md:
          type: string
          description: Content to inject as CLAUDE.md into the agent container.
          example: "# Project instructions\n..."
        env_vars:
          type: object
          additionalProperties:
            type: string
          description: Additional environment variables to set in the agent container.
          example:
            CUSTOM_SETTING: "value"
...